+++ src/pages/Main.tsx
@@ -1,41 +1,8 @@
-import { useEffect, useState } from "react";
-import { useSetRecoilState } from "recoil";
-import { fetchCartItems, fetchProducts } from "../api";
 import { Header, Loading, Page, ProductList } from "../components";
-import { MIN_QUANTITY } from "../constants";
-import { productsState } from "../recoil/atom";
-import { CartItemType, PayloadType } from "../types/domain";
+import { useFetch } from "../hooks/useFetch";
 
 const Main = () => {
-  const setProducts = useSetRecoilState(productsState);
-  const [isLoading, setIsLoading] = useState(true);
-
-  useEffect(() => {
-    setProductsWithQuantity();
-  }, []);
-
-  const setProductsWithQuantity = async () => {
-    try {
-      const products = await fetchProducts();
-      const cartItems = await fetchCartItems();
-
-      const productsWithQuantity = products.map((product: PayloadType) => {
-        const cartProduct = cartItems.find(
-          (cartItem: CartItemType) => cartItem.id === product.id
-        );
-        return {
-          ...product,
-          quantity: cartProduct
-            ? cartProduct.quantity
-            : MIN_QUANTITY.toString(),
-        };
-      });
-      setProducts(productsWithQuantity);
-      setIsLoading(false);
-    } catch (error) {
-      console.log(error);
-    }
-  };
+  const { isLoading } = useFetch();
 
   return (
     <>
